Bug Fixes and TravisCI Integration
- Added deploy hook for TravisCI
- Fixed .isDirectory() not a function in get-spec.ts issue
- Reorganized declarations of types
