--- .npmignore
@@ -1,5 +1,0 @@
-src
-docs
-test
-coverage
-.nyc_output
