+++ test/index.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import 'mocha';
-import * as index from '../src/index';
+import * as index from '../dist/index';
 
 describe('index', () => {
   it('should have apply function.', () => {
